refactor: move snippets to proper dir and format code
