+++ python/eggroll/computing/tasks/impl/_binary_sorted_map_partitions_with_index.py
@@ -3,7 +3,7 @@ import logging
 import typing
 from typing import Callable, Iterable
 
-from eggroll.computing.tasks import consts, store
+from eggroll.computing.tasks import consts, store, job_util
 from eggroll.computing.tasks.submit_utils import block_submit_unary_unit_job
 from eggroll.core.meta_model import (
     ErJob,
@@ -11,7 +11,6 @@ from eggroll.core.meta_model import (
     ErTask,
     ErJobIO,
 )
-from eggroll.core.utils import generate_job_id
 from ._task import Task, EnvOptions
 
 if typing.TYPE_CHECKING:
@@ -67,7 +66,7 @@ class BinarySortedMapPartitionsWithIndex(Task):
             func=func,
         )
         job = ErJob(
-            id=generate_job_id(
+            id=job_util.generate_job_id(
                 left.session_id, consts.BINARY_SORTED_MAP_PARTITIONS_WITH_INDEX
             ),
             name=consts.BINARY_SORTED_MAP_PARTITIONS_WITH_INDEX,
