+++ python/eggroll/computing/tasks/impl/_map_partitions_with_index.py
@@ -3,9 +3,8 @@ import logging
 import typing
 import uuid
 
-from eggroll.computing.tasks import consts, store
+from eggroll.computing.tasks import consts, store, job_util, transfer
 from eggroll.computing.tasks.submit_utils import block_submit_unary_unit_job
-from eggroll.computing.tasks.transfer_pair import BatchBroker, TransferPair
 from eggroll.core.datastructure.broker import FifoBroker
 from eggroll.core.meta_model import (
     ErJob,
@@ -16,7 +15,6 @@ from eggroll.core.meta_model import (
     ErSerdes,
     MapPartitionsWithIndexRequest,
 )
-from eggroll.core.utils import generate_job_id
 from ._task import Task, EnvOptions
 
 if typing.TYPE_CHECKING:
@@ -87,7 +85,9 @@ class MapReducePartitionsWithIndex(Task):
             name=f"{consts.MAP_REDUCE_PARTITIONS_WITH_INDEX}_{shuffle}",
             body=MapPartitionsWithIndexRequest(shuffle=shuffle).to_proto_string(),
         )
-        job_id = generate_job_id(rp.session_id, consts.MAP_REDUCE_PARTITIONS_WITH_INDEX)
+        job_id = job_util.generate_job_id(
+            rp.session_id, consts.MAP_REDUCE_PARTITIONS_WITH_INDEX
+        )
         output_store = rp.ctx.create_store(
             id=rp.get_store().store_locator.id,
             name=output_name,
@@ -158,7 +158,7 @@ class MapReducePartitionsWithIndex(Task):
             or task.first_output.is_on_node(env_options.server_node_id)
         )
 
-        shuffler = TransferPair(config=env_options.config, transfer_id=job.id)
+        shuffler = transfer.TransferPair(config=env_options.config, transfer_id=job.id)
         store_broker_future = None
         if task_has_output:
             task_output = task.first_output
@@ -195,7 +195,9 @@ class MapReducePartitionsWithIndex(Task):
                     ).iteritems()
                 )
                 task_shuffle_write_batch_broker = stack.enter_context(
-                    BatchBroker(config=env_options.config, broker=shuffle_write_broker)
+                    transfer.BatchBroker(
+                        config=env_options.config, broker=shuffle_write_broker
+                    )
                 )
                 partition_id = task.first_input.id
                 value = map_op(partition_id, task_input_iterator)
