+++ python/eggroll/computing/tasks/store/_adapter.py
