+++ python/eggroll/computing/tasks/store/_format.py
