+++ python/eggroll/computing/tasks/store/_type.py
@@ -1,0 +1,3 @@
+class StoreTypes(object):
+    ROLLPAIR_IN_MEMORY = "IN_MEMORY"
+    ROLLPAIR_LMDB = "LMDB"
