+++ python/eggroll/core/command/command_status.py
@@ -1,0 +1,7 @@
+class SessionStatus(object):
+    NEW = "NEW"
+    ACTIVE = "ACTIVE"
+    CLOSED = "CLOSED"
+    KILLED = "KILLED"
+    ERROR = "ERROR"
+    FINISHED = "FINISHED"
