+++ python/eggroll/core/command/command_uri.py
