+++ python/eggroll/core/constants.py
@@ -13,62 +13,6 @@
 #  limitations under the License.
 
 
-class ServerNodeStatus(object):
-    HEALTHY = "HEALTHY"
-
-
-class ServerNodeTypes(object):
-    CLUSTER_MANAGER = "CLUSTER_MANAGER"
-    NODE_MANAGER = "NODE_MANAGER"
-
-
-class StoreTypes(object):
-    ROLLPAIR_IN_MEMORY = "IN_MEMORY"
-    ROLLPAIR_LMDB = "LMDB"
-    ROLLPAIR_LEVELDB = "LEVELDB"
-    ROLLFRAME_FILE = "ROLL_FRAME_FILE"
-    ROLLPAIR_ROLLSITE = "ROLL_SITE"
-    ROLLPAIR_FILE = "ROLL_PAIR_FILE"
-    ROLLPAIR_MMAP = "ROLL_PAIR_MMAP"
-    ROLLPAIR_CACHE = "ROLL_PAIR_CACHE"
-    ROLLPAIR_QUEUE = "ROLL_PAIR_QUEUE"
-
-
-class PartitionerTypes(object):
-    MMH3_BYTESTRING_HASH = "MMH3_BYTESTRING_HASH"
-    INTEGER_HASH = "INTEGER_KEY_HASH"
-
-
 class ProcessorTypes(object):
     EGG_PAIR = "egg_pair"
     ROLL_PAIR_MASTER = "roll_pair_master"
-
-
-class RollTypes(object):
-    ROLL_PAIR = "roll_pair"
-
-
-class ProcessorStatus(object):
-    NEW = "NEW"
-    RUNNING = "RUNNING"
-    STOPPED = "STOPPED"
-    KILLED = "KILLED"
-
-
-class SessionStatus(object):
-    NEW = "NEW"
-    ACTIVE = "ACTIVE"
-    CLOSED = "CLOSED"
-    KILLED = "KILLED"
-    ERROR = "ERROR"
-    FINISHED = "FINISHED"
-
-
-class DeployModes(object):
-    STANDALONE = "standalone"
-    CLUSTER = "cluster"
-
-
-class Charsets(object):
-    UTF8 = "utf8"
-    ISO_8859_1 = "iso-8859-1"
