+++ python/eggroll/trace/__init__.py
@@ -1,3 +1,6 @@
+from ._egg_pair import get_system_metric
+from ._exception_logger import exception_catch
+from ._roll_pair import roll_pair_method_trace
 from ._tracer import (
     inject_carrier,
     setup_tracing,
@@ -5,9 +8,6 @@ from ._tracer import (
     extract_carrier,
     get_tracer,
 )
-from ._roll_pair import roll_pair_method_trace
-from ._egg_pair import get_system_metric
-
 
 __all__ = [
     "setup_tracing",
@@ -17,4 +17,5 @@ __all__ = [
     "get_tracer",
     "roll_pair_method_trace",
     "get_system_metric",
+    "exception_catch",
 ]
